chore: rename bearer output to saas
